refactor(ckbtc-minter): move ICRC-21 tests into updates/tests.rs
Group all updates-module tests in the same file (updates/tests.rs) the
way update_balance is already organized: keep production code in
updates/icrc21.rs and put the tests in a sibling `mod icrc21 {...}`
block in updates/tests.rs.

Bumps the visibility of the items the tests reach into — DECIMALS,
TokenSymbols, TokenSymbols::for_network, format_amount, and
build_consent_info — to pub(super), which is the minimum needed for
the sibling `crate::updates::tests::icrc21` module to use them while
keeping them out of the crate-public API.
