+++ rs/bitcoin/ckbtc/minter/src/updates/tests.rs
@@ -1164,3 +1164,315 @@ mod update_balance {
         CkBtcEventLogger.events_iter()
     }
 }
+
+mod icrc21 {
+    use crate::Network;
+    use crate::updates::icrc21::{DECIMALS, TokenSymbols, build_consent_info, format_amount};
+    use crate::updates::retrieve_btc::RetrieveBtcWithApprovalArgs;
+    use candid::Encode;
+    use icrc_ledger_types::icrc21::errors::Icrc21Error;
+    use icrc_ledger_types::icrc21::lib::MAX_CONSENT_MESSAGE_ARG_SIZE_BYTES;
+    use icrc_ledger_types::icrc21::requests::{
+        ConsentMessageMetadata, ConsentMessageRequest, ConsentMessageSpec, DisplayMessageType,
+    };
+    use icrc_ledger_types::icrc21::responses::{ConsentMessage, Value};
+
+    fn make_request(
+        method: &str,
+        arg: Vec<u8>,
+        device_spec: Option<DisplayMessageType>,
+    ) -> ConsentMessageRequest {
+        ConsentMessageRequest {
+            method: method.to_string(),
+            arg,
+            user_preferences: ConsentMessageSpec {
+                metadata: ConsentMessageMetadata {
+                    language: "en".to_string(),
+                    utc_offset_minutes: None,
+                },
+                device_spec,
+            },
+        }
+    }
+
+    #[test]
+    fn test_format_amount() {
+        assert_eq!(format_amount(0, 8), "0");
+        assert_eq!(format_amount(1, 8), "0.00000001");
+        assert_eq!(format_amount(100_000_000, 8), "1");
+        assert_eq!(format_amount(150_000_000, 8), "1.5");
+        assert_eq!(format_amount(123_456_789, 8), "1.23456789");
+    }
+
+    #[test]
+    fn test_unsupported_method() {
+        // Includes `retrieve_btc` because the minter intentionally only
+        // supports ICRC-21 for the approval-based flow — wallets calling
+        // `retrieve_btc` should not get a consent message rendered for them.
+        for method in ["update_balance", "retrieve_btc", "get_btc_address", ""] {
+            let req = make_request(method, vec![], None);
+            let err = build_consent_info(req, Network::Mainnet).unwrap_err();
+            assert!(
+                matches!(err, Icrc21Error::UnsupportedCanisterCall(_)),
+                "method {method:?} should be unsupported, got {err:?}"
+            );
+        }
+    }
+
+    #[test]
+    fn test_argument_too_large() {
+        let req = make_request(
+            "retrieve_btc_with_approval",
+            vec![0; MAX_CONSENT_MESSAGE_ARG_SIZE_BYTES as usize + 1],
+            None,
+        );
+        let err = build_consent_info(req, Network::Mainnet).unwrap_err();
+        match err {
+            Icrc21Error::UnsupportedCanisterCall(info) => {
+                assert!(info.description.contains("argument size is too large"));
+            }
+            _ => panic!("expected UnsupportedCanisterCall, got {err:?}"),
+        }
+    }
+
+    #[test]
+    fn test_retrieve_btc_with_approval_generic_display() {
+        let args = RetrieveBtcWithApprovalArgs {
+            amount: 150_000,
+            address: "bc1qar0srrr7xfkvy5l643lydnw9re59gtzzwf5mdq".to_string(),
+            from_subaccount: None,
+        };
+        let req = make_request(
+            "retrieve_btc_with_approval",
+            Encode!(&args).unwrap(),
+            Some(DisplayMessageType::GenericDisplay),
+        );
+        let info = build_consent_info(req, Network::Mainnet).unwrap();
+        assert_eq!(info.metadata.language, "en");
+        let message = match info.consent_message {
+            ConsentMessage::GenericDisplayMessage(m) => m,
+            other => panic!("expected GenericDisplayMessage, got {other:?}"),
+        };
+        assert!(message.starts_with("# Convert ckBTC to BTC"));
+        assert!(message.contains("0.0015 ckBTC"));
+        assert!(message.contains("bc1qar0srrr7xfkvy5l643lydnw9re59gtzzwf5mdq"));
+        // No subaccount section if from_subaccount is None.
+        assert!(!message.contains("source subaccount"));
+    }
+
+    #[test]
+    fn test_retrieve_btc_with_approval_generic_display_with_subaccount() {
+        let mut subaccount = [0_u8; 32];
+        subaccount[31] = 0x42;
+        let args = RetrieveBtcWithApprovalArgs {
+            amount: 100_000_000,
+            address: "bc1qar0srrr7xfkvy5l643lydnw9re59gtzzwf5mdq".to_string(),
+            from_subaccount: Some(subaccount),
+        };
+        let req = make_request("retrieve_btc_with_approval", Encode!(&args).unwrap(), None);
+        let info = build_consent_info(req, Network::Mainnet).unwrap();
+        let message = match info.consent_message {
+            ConsentMessage::GenericDisplayMessage(m) => m,
+            other => panic!("expected GenericDisplayMessage, got {other:?}"),
+        };
+        assert!(message.contains("1 ckBTC"));
+        assert!(message.contains(&hex::encode(subaccount)));
+    }
+
+    #[test]
+    fn test_retrieve_btc_with_approval_fields_display() {
+        // Long values (Bitcoin address, subaccount hex) are emitted as a
+        // single Value::Text — wallets paginate them across screens. The
+        // number of fields is therefore independent of the value length.
+        let mut subaccount = [0_u8; 32];
+        subaccount[0] = 0xab;
+        subaccount[31] = 0xcd;
+        let address = "bc1qar0srrr7xfkvy5l643lydnw9re59gtzzwf5mdq".to_string();
+        let args = RetrieveBtcWithApprovalArgs {
+            amount: 250_000,
+            address: address.clone(),
+            from_subaccount: Some(subaccount),
+        };
+        let req = make_request(
+            "retrieve_btc_with_approval",
+            Encode!(&args).unwrap(),
+            Some(DisplayMessageType::FieldsDisplay),
+        );
+        let info = build_consent_info(req, Network::Mainnet).unwrap();
+        let fields_display = match info.consent_message {
+            ConsentMessage::FieldsDisplayMessage(f) => f,
+            other => panic!("expected FieldsDisplayMessage, got {other:?}"),
+        };
+        assert_eq!(fields_display.intent, "ckBTC to BTC");
+        assert_eq!(
+            fields_display.fields,
+            vec![
+                (
+                    "Amount".to_string(),
+                    Value::TokenAmount {
+                        decimals: DECIMALS,
+                        amount: 250_000,
+                        symbol: "ckBTC".to_string(),
+                    }
+                ),
+                ("BTC address".to_string(), Value::Text { content: address }),
+                (
+                    "From subaccount".to_string(),
+                    Value::Text {
+                        content: hex::encode(subaccount)
+                    }
+                ),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_retrieve_btc_with_approval_fields_display_no_subaccount() {
+        let args = RetrieveBtcWithApprovalArgs {
+            amount: 250_000,
+            address: "bc1qar0srrr7xfkvy5l643lydnw9re59gtzzwf5mdq".to_string(),
+            from_subaccount: None,
+        };
+        let req = make_request(
+            "retrieve_btc_with_approval",
+            Encode!(&args).unwrap(),
+            Some(DisplayMessageType::FieldsDisplay),
+        );
+        let info = build_consent_info(req, Network::Mainnet).unwrap();
+        let fields_display = match info.consent_message {
+            ConsentMessage::FieldsDisplayMessage(f) => f,
+            other => panic!("expected FieldsDisplayMessage, got {other:?}"),
+        };
+        assert_eq!(fields_display.fields.len(), 2);
+        // No subaccount field when from_subaccount is None.
+        assert!(
+            !fields_display
+                .fields
+                .iter()
+                .any(|(label, _)| label == "From subaccount")
+        );
+    }
+
+    #[test]
+    fn test_token_symbols_for_network() {
+        assert_eq!(
+            TokenSymbols::for_network(Network::Mainnet),
+            TokenSymbols {
+                ckbtc: "ckBTC",
+                btc: "BTC"
+            }
+        );
+        assert_eq!(
+            TokenSymbols::for_network(Network::Testnet),
+            TokenSymbols {
+                ckbtc: "ckTESTBTC",
+                btc: "TESTBTC"
+            }
+        );
+        assert_eq!(
+            TokenSymbols::for_network(Network::Regtest),
+            TokenSymbols {
+                ckbtc: "ckTESTBTC",
+                btc: "TESTBTC"
+            }
+        );
+    }
+
+    #[test]
+    fn test_retrieve_btc_with_approval_uses_testnet_symbols() {
+        let args = RetrieveBtcWithApprovalArgs {
+            amount: 250_000,
+            address: "tb1qw508d6qejxtdg4y5r3zarvary0c5xw7kxpjzsx".to_string(),
+            from_subaccount: None,
+        };
+        let req = make_request(
+            "retrieve_btc_with_approval",
+            Encode!(&args).unwrap(),
+            Some(DisplayMessageType::FieldsDisplay),
+        );
+        let info = build_consent_info(req, Network::Testnet).unwrap();
+        let fields_display = match info.consent_message {
+            ConsentMessage::FieldsDisplayMessage(f) => f,
+            other => panic!("expected FieldsDisplayMessage, got {other:?}"),
+        };
+        assert_eq!(fields_display.intent, "ckTESTBTC to TESTBTC");
+        match &fields_display.fields[0].1 {
+            Value::TokenAmount { symbol, .. } => assert_eq!(symbol, "ckTESTBTC"),
+            other => panic!("expected TokenAmount, got {other:?}"),
+        }
+        // FieldsDisplay address label uses the testnet native symbol too.
+        assert_eq!(fields_display.fields[1].0, "TESTBTC address");
+    }
+
+    #[test]
+    fn test_retrieve_btc_with_approval_generic_uses_testnet_symbols() {
+        let args = RetrieveBtcWithApprovalArgs {
+            amount: 100_000_000,
+            address: "tb1qw508d6qejxtdg4y5r3zarvary0c5xw7kxpjzsx".to_string(),
+            from_subaccount: None,
+        };
+        let req = make_request(
+            "retrieve_btc_with_approval",
+            Encode!(&args).unwrap(),
+            Some(DisplayMessageType::GenericDisplay),
+        );
+        let info = build_consent_info(req, Network::Testnet).unwrap();
+        let message = match info.consent_message {
+            ConsentMessage::GenericDisplayMessage(m) => m,
+            other => panic!("expected GenericDisplayMessage, got {other:?}"),
+        };
+        assert!(message.starts_with("# Convert ckTESTBTC to TESTBTC"));
+        assert!(message.contains("1 ckTESTBTC"));
+        assert!(message.contains("ckTESTBTC minter"));
+        assert!(message.contains("equivalent amount in TESTBTC"));
+    }
+
+    #[test]
+    fn test_invalid_args() {
+        let req = make_request("retrieve_btc_with_approval", vec![1, 2, 3], None);
+        let err = build_consent_info(req, Network::Mainnet).unwrap_err();
+        match err {
+            Icrc21Error::UnsupportedCanisterCall(info) => {
+                assert!(info.description.contains("Failed to decode"));
+            }
+            _ => panic!("expected UnsupportedCanisterCall, got {err:?}"),
+        }
+    }
+
+    #[test]
+    fn test_malformed_address_is_rejected() {
+        // The minter must not interpolate an unparseable address into the
+        // Markdown consent message — that would be a Markdown-injection vector
+        // (e.g. an "address" containing newlines, backticks, or '#' that fakes
+        // additional fields).
+        for bad_address in [
+            "not-a-real-address",
+            "bc1qar0srrr7xfkvy5l643lydnw9re59gtzzwf5mdq\n# You will receive 100 BTC",
+            "1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa`\n\n**Amount:** 100 BTC\n`",
+            "tb1qw508d6qejxtdg4y5r3zarvary0c5xw7kxpjzsx", // valid testnet but on Mainnet
+        ] {
+            let args = RetrieveBtcWithApprovalArgs {
+                amount: 50_000,
+                address: bad_address.to_string(),
+                from_subaccount: None,
+            };
+            let req = make_request(
+                "retrieve_btc_with_approval",
+                Encode!(&args).unwrap(),
+                Some(DisplayMessageType::GenericDisplay),
+            );
+            let err = build_consent_info(req, Network::Mainnet).unwrap_err();
+            match err {
+                Icrc21Error::UnsupportedCanisterCall(info) => {
+                    assert!(
+                        info.description
+                            .contains("Invalid Bitcoin destination address"),
+                        "unexpected error description: {}",
+                        info.description
+                    );
+                }
+                other => panic!("expected UnsupportedCanisterCall, got {other:?}"),
+            }
+        }
+    }
+}
